Switcher move _async_call_api to entity.py
